requirements.txt
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
